+++ src/coq_elpi_utils.mli
@@ -7,6 +7,8 @@ val to_coq_loc : Elpi.API.Ast.Loc.t -> Loc.t
 val of_coq_loc : Loc.t -> Elpi.API.Ast.Loc.t
 
 val err : ?loc:Elpi.API.Ast.Loc.t -> Pp.t -> 'a
+exception FatalElpiError of Pp.t
+val fatal_err : ?loc:Elpi.API.Ast.Loc.t -> Pp.t -> 'a
 val nYI : string -> 'a
 val safe_destApp : Evd.evar_map ->
   EConstr.t -> (EConstr.t,EConstr.types,EConstr.ESorts.t, EConstr.EInstance.t) Constr.kind_of_term * EConstr.t array